Added Delete confirm to all modules

+++ src/app/components/mobileTags/mobile-tags.component.ts
@@ -1,56 +1,64 @@
-import { Component, OnInit, ViewChild, AfterViewInit } from '@angular/core';
-import { Router, ActivatedRoute } from '@angular/router';
-import { MobileTags } from '../../models/mobile-tags';
-import { MatTableDataSource } from '@angular/material/table';
-import { MobileTagsService } from '../../../app/services/mobile-tags.service';
-import { startWith, tap } from 'rxjs/operators';
-import { MatPaginator } from '@angular/material';
+import { Component, OnInit, ViewChild, AfterViewInit } from "@angular/core";
+import { Router, ActivatedRoute } from "@angular/router";
+import { MobileTags } from "../../models/mobile-tags";
+import { MatTableDataSource } from "@angular/material/table";
+import { MobileTagsService } from "../../../app/services/mobile-tags.service";
+import { startWith, tap } from "rxjs/operators";
+import { MatPaginator } from "@angular/material";
 
 export class mobileTagFilter {
-  name: string = '';
+  name: string = "";
 }
 
 @Component({
-  selector: 'app-mobile-tags',
-  templateUrl: './mobile-tags.component.html',
-  styleUrls: ['./mobile-tags.component.css']
+  selector: "app-mobile-tags",
+  templateUrl: "./mobile-tags.component.html",
+  styleUrls: ["./mobile-tags.component.css"]
 })
-
 export class MobileTagsComponent implements OnInit, AfterViewInit {
-
   ngAfterViewInit(): void {
     // let pageIndex = this.paginator.pageIndex + 1
 
-    this.paginator.page.pipe(
-      startWith(null),
-      tap(() => this.getTags(this.paginator.pageIndex+1, this.paginator.pageSize))).subscribe();
+    this.paginator.page
+      .pipe(
+        startWith(null),
+        tap(() =>
+          this.getTags(this.paginator.pageIndex + 1, this.paginator.pageSize)
+        )
+      )
+      .subscribe();
   }
   @ViewChild(MatPaginator, { static: false })
-  paginator: MatPaginator
+  paginator: MatPaginator;
 
-  constructor(private router: Router,
-    private activatedRoute: ActivatedRoute, private tagService: MobileTagsService) { }
+  constructor(
+    private router: Router,
+    private activatedRoute: ActivatedRoute,
+    private tagService: MobileTagsService
+  ) {}
   TagModel: MobileTags = new MobileTags();
-  displayedColumns: string[] = ['id', 'name', 'type', 'action'];
+  displayedColumns: string[] = ["id", "name", "type", "action"];
   dataSource = new MatTableDataSource<any>([]);
-  count: number
+  count: number;
   /*Table logic*/
   deleteCategory(row) {
-    this.tagService.delete(row._id).subscribe((result: any) => {
-      if (result.status == 200) {
-        this.getTags(1, 5);
-        // this.dataSource.data
-      }
-
-    })
+    if (confirm("Are you sure to delete this Tag?")) {
+      this.tagService.delete(row._id).subscribe((result: any) => {
+        if (result.status == 200) {
+          this.getTagCount();
+          this.getTags(this.paginator.pageIndex + 1, this.paginator.pageSize);
+          // this.dataSource.data
+        }
+      });
+    }
   }
 
   routeToTagForm() {
-    this.router.navigate(['MobileTagForm'], { queryParams: { id: 'new' } });
+    this.router.navigate(["MobileTagForm"], { queryParams: { id: "new" } });
   }
 
   editTag(row) {
-    this.router.navigate(['home/MobileTagForm'], row._id);
+    this.router.navigate(["home/MobileTagForm"], row._id);
   }
 
   getTags(pageIndex, size) {
@@ -58,8 +66,7 @@ export class MobileTagsComponent implements OnInit, AfterViewInit {
       if (result.status == 200) {
         this.dataSource = new MatTableDataSource<any>(result.data);
       }
-
-    })
+    });
   }
   applyFilter(filterValue: string) {
     this.dataSource.filter = filterValue.trim().toLowerCase();
@@ -80,7 +87,6 @@ export class MobileTagsComponent implements OnInit, AfterViewInit {
       if (result.success) {
         this.count = result.count;
       }
-    })
+    });
   }
-
 }
